mover metodo ask_set_content a xbmc_library
